Added example for “double” proxying.
If Spring Data Repositories are used in combination with @GlobalMethodSecurity with proxyTargetClass=true and @Pre/Post Security annotations are used on repository finder methods, then Spring framework tries to create a cglib-proxyclass for the repository instance which is a JDK Proxy.

This is a problem that puzzles a lot of beginners.

This project contains a shadowed version of AbstractAutoProxyCreator that is able to detect an already Avoided target instance and instead of generating a new class - simply registers the additional advise.